+++ src/hooks/use100vh.js
@@ -1,0 +1,17 @@
+import React, { useEffect } from "react";
+
+function use100vh() {
+  const handleResize = () => {
+    const vh = window.innerHeight * 0.01;
+    document.documentElement.style.setProperty("--vh", `${vh}px`);
+  };
+
+  useEffect(() => {
+    window.addEventListener("resize", handleResize);
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, []);
+}
+
+export default use100vh;
